backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
dj-database-url==2.1.0
anthropic==0.25.0
python-dotenv==1.0.0
